+++ src/main/java/com/banco/domain/conta/exception/ContaNaoEncontradaException.java
@@ -4,7 +4,7 @@ public class ContaNaoEncontradaException extends RuntimeException {
 
     private static final String MENSAGEM_PADRAO = "Conta de número %d não encontrada";
 
-    public ContaNaoEncontradaException(int numeroConta) {
+    public ContaNaoEncontradaException(long numeroConta) {
         super(String.format(MENSAGEM_PADRAO, numeroConta));
     }
